Split hosted review creation checks
