+++ config/max-lines-baseline.txt
@@ -26,7 +26,6 @@ inline src/main/runtime/orca-runtime.test.ts
 inline src/main/runtime/orca-runtime.ts
 inline src/main/runtime/rpc/methods/orchestration.ts
 inline src/main/runtime/runtime-rpc.ts
-inline src/main/source-control/hosted-review-creation.ts
 inline src/main/ssh/ssh-channel-multiplexer.ts
 inline src/main/ssh/ssh-connection.ts
 inline src/main/ssh/ssh-relay-deploy.ts
